+++ admin/0.BUILD_PACKAGE.R
@@ -1,20 +1,41 @@
-devtools::document()
-devtools::load_all()
+# 
 
-devtools::build()
+# Prepare files -----------------------------------------------------------
 
-devtools::check()
+  # Copy canonical_protocol/tasks to inst/templates/tasks.zip
+  files_tasks = list.files(path = "canonical_protocol/tasks", full.names = TRUE, recursive = TRUE)
+  zip(zipfile = "inst/templates/tasks.zip", files = files_tasks, extras = "-j")
 
-devtools::install()
+  # Copy canonical_protocol_clean to inst/templates/canonical_protocol_clean.zip
+  setwd("canonical_protocol_clean/")
+  files_canonical_protocol_clean = list.files(path = ".", full.names = TRUE, recursive = TRUE)
+  zip(zipfile = "../inst/templates/canonical_protocol_clean.zip", files = files_canonical_protocol_clean)
+  setwd("../")
+  
 
-# TODO: As a part of the build package process, should COPY "canonical_protocol_clean" to "inst/templates/canonical_protocol_clean.zip"
+# Build package -----------------------------------------------------------
+
+  devtools::document()
+  devtools::load_all()
+  devtools::build()
+  
+  devtools::check()
+  
+  devtools::install()
+
+
+# TODO --------------------------------------------------------------------
+
+
 # TODO: Add more template plugins!
+# TODO: tasks with images
+# TODO: shiny app to modify local config.js? Select folder...
 
-# 1 warning left: 
-  # checking R files for non-ASCII characters ...
-  # Found the following file with non-ASCII characters:
-  #   create_task.R
-  # Portable packages must use only ASCII characters in their R code,
-  # except perhaps in comments.
-  # Use \uxxxx escapes for other characters.
 
+
+# REMEMBER ----------------------------------------------------------------
+
+# If warning about non-ASCII characters. Find character and replace 
+  # https://altcodeunicode.com/alt-codes-letter-o-with-accents/
+  # tools::showNonASCIIfile(file = "R/create_task.R")
+  # e.g. ó -> \u00F3